feat: prepare secure customer launch

+++ backend/tests/tokens.rs
@@ -2,6 +2,7 @@ use alpha_studio_backend::tokens::{
     AdminTokenClaims, AdminTokenService, DeviceTokenClaims, DeviceTokenService, RunTokenClaims,
     RunTokenService,
 };
+use rust_decimal::Decimal;
 
 #[test]
 fn run_tokens_bind_tenant_device_model_and_budget() {
@@ -13,7 +14,7 @@ fn run_tokens_bind_tenant_device_model_and_budget() {
             "device_1".to_string(),
             "run_1".to_string(),
             "gpt-5.5".to_string(),
-            10.0,
+            Decimal::from(10_u64),
             60,
         ))
         .expect("token should be issued");
@@ -23,7 +24,7 @@ fn run_tokens_bind_tenant_device_model_and_budget() {
     assert_eq!(claims.tenant_id, "tenant_1");
     assert_eq!(claims.device_id, "device_1");
     assert_eq!(claims.model_id, "gpt-5.5");
-    assert_eq!(claims.budget_yuan, 10.0);
+    assert_eq!(claims.budget_yuan, Decimal::from(10_u64));
 }
 
 #[test]
